chore(risk): share judge message payload rendering
Move the structured judge message payload renderer into judgemessage so promptinjection no longer depends on promptpolicy/openrouter. Keep prompt policy and prompt injection dependencies as function types.
