+++ server/cmd/risk-pi-report/main.go
@@ -28,7 +28,6 @@ import (
 	"github.com/speakeasy-api/gram/server/internal/scanners"
 	"github.com/speakeasy-api/gram/server/internal/scanners/promptinjection"
 	piopenrouter "github.com/speakeasy-api/gram/server/internal/scanners/promptinjection/openrouter"
-	ppopenrouter "github.com/speakeasy-api/gram/server/internal/scanners/promptpolicy/openrouter"
 	"github.com/speakeasy-api/gram/server/internal/thirdparty/openrouter"
 )
 
@@ -624,13 +623,13 @@ func scanJudge(ctx context.Context, opts options, client openrouter.CompletionCl
 }
 
 // judgeOne issues one GetObjectCompletion shaped exactly like piopenrouter's call:
-// the structured "message" payload (ppopenrouter.RenderMessage), piopenrouter's system
-// prompt and verdict schema, temperature 0. No copy of the prompt/schema to keep
-// in sync — it drives the production constants directly.
+// the structured message payload, piopenrouter's system prompt and verdict
+// schema, temperature 0. No copy of the prompt/schema to keep in sync - it
+// drives the production constants directly.
 func judgeOne(ctx context.Context, client openrouter.CompletionClient, model string, msg judgemessage.Message) (isAttack bool, confidence float64, err error) {
 	payload, err := json.Marshal(struct {
-		Message ppopenrouter.MessagePayload `json:"message"`
-	}{Message: ppopenrouter.RenderMessage(msg)})
+		Message judgemessage.Payload `json:"message"`
+	}{Message: judgemessage.RenderPayload(msg)})
 	if err != nil {
 		return false, 0, fmt.Errorf("marshal judge payload: %w", err)
 	}
