+++ server/internal/background/activities/risk_analysis/analyze_batch_test.go
@@ -344,7 +344,7 @@ func TestAnalyzeBatch_PromptJudgeUsesToolCallPayload(t *testing.T) {
 		nil,
 		nil,
 		nil,
-		judge,
+		judge.Evaluate,
 		flags,
 		newPresidioPub(),
 		newGitleaksPub(),
@@ -430,7 +430,7 @@ func TestAnalyzeBatch_PromptJudgeMultiToolCallAttribution(t *testing.T) {
 		nil,
 		nil,
 		nil,
-		judge,
+		judge.Evaluate,
 		flags,
 		newPresidioPub(),
 		newGitleaksPub(),
