+++ server/internal/scanners/promptinjection/scanner.go
@@ -32,27 +32,15 @@ type Result struct {
 	Rationale string
 }
 
-type Classifier interface {
-	Classify(ctx context.Context, req Request) ([]Result, error)
-}
+type Classifier func(ctx context.Context, req Request) ([]Result, error)
 
-type ClassifierFunc func(ctx context.Context, req Request) ([]Result, error)
-
-func (f ClassifierFunc) Classify(ctx context.Context, req Request) ([]Result, error) {
-	return f(ctx, req)
-}
-
-type noopClassifier struct{}
-
-func (noopClassifier) Classify(_ context.Context, req Request) ([]Result, error) {
+func NoopClassifier(_ context.Context, req Request) ([]Result, error) {
 	results := make([]Result, len(req.Messages))
 	for i := range results {
 		results[i] = Result{Label: LabelSafe, Score: 0, Rationale: ""}
 	}
 	return results, nil
 }
-
-var NoopClassifier Classifier = noopClassifier{}
 
 func Describe() (string, string) {
 	return scanners.GuardRuleID(Rule), "Detected a prompt injection attempt."
@@ -85,7 +73,7 @@ func (s *Scanner) Scan(ctx context.Context, text, orgID, projectID string, msg j
 		return findings, nil
 	}
 
-	results, err := s.classifier.Classify(ctx, Request{Messages: []judgemessage.Message{msg}, OrgID: orgID, ProjectID: projectID})
+	results, err := s.classifier(ctx, Request{Messages: []judgemessage.Message{msg}, OrgID: orgID, ProjectID: projectID})
 	if err != nil {
 		s.logger.WarnContext(ctx, "pi L1 scan failed; returning L0 findings only",
 			attr.SlogError(err),
@@ -116,7 +104,7 @@ func (s *Scanner) ScanBatch(ctx context.Context, texts []string, orgID, projectI
 		return out, nil
 	}
 
-	results, err := s.classifier.Classify(ctx, Request{Messages: msgs, OrgID: orgID, ProjectID: projectID})
+	results, err := s.classifier(ctx, Request{Messages: msgs, OrgID: orgID, ProjectID: projectID})
 	if err != nil {
 		s.logger.WarnContext(ctx, "pi L1 batch scan failed; returning L0 findings only",
 			attr.SlogError(err),
